+++ src/steps/resource-manager/authorization/converters.test.ts
@@ -1,12 +1,11 @@
-import { convertProperties, Entity } from '@jupiterone/integration-sdk-core';
+import { convertProperties } from '@jupiterone/integration-sdk-core';
 
 import { createAzureWebLinker } from '../../../azure';
 import {
   createRoleDefinitionEntity,
-  createRoleAssignmentDirectRelationship,
-  createRoleAssignmentMappedRelationship,
   createClassicAdministratorEntity,
   createClassicAdministratorHasUserRelationship,
+  createRoleAssignmentEntity,
 } from './converters';
 import {
   RoleDefinition,
@@ -48,7 +47,7 @@ describe('createRoleDefinitionEntity', () => {
       _key:
         '/providers/Microsoft.Authorization/roleDefinitions/8e3af657-a8ff-443c-a75c-2fe8c4bcb635',
       _type: 'azure_role_definition',
-      _class: ['AccessRole', 'AccessPolicy'],
+      _class: ['AccessPolicy'],
       _rawData: [{ name: 'default', rawData: data }],
       displayName: 'Owner',
       type: 'Microsoft.Authorization/roleDefinitions',
@@ -63,118 +62,55 @@ describe('createRoleDefinitionEntity', () => {
   });
 });
 
-describe('createRoleAssignmentDirectRelationship', () => {
+describe('createRoleAssignmentEntity', () => {
   test('properties transferred', () => {
-    const roleAssignment: RoleAssignment = {
+    const data: RoleAssignment = {
       id:
-        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleAssignments/c967042a-aad6-4e3b-8485-1c85d5e6f9e8',
-      name: 'c967042a-aad6-4e3b-8485-1c85d5e6f9e8',
+        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleAssignments/10000000-0000-0000-0000-000000000000',
+      name: '10000000-0000-0000-0000-000000000000',
       type: 'Microsoft.Authorization/roleAssignments',
       scope: '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7',
       roleDefinitionId:
-        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleDefinitions/8e3af657-a8ff-443c-a75c-2fe8c4bcb635',
-      principalId: 'b6e7f627-8731-47bd-be0e-477d1cbc6e17',
-      principalType: 'User',
+        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleDefinitions/8bfcfc94-cf28-d595-8e3d-851a1eb7c8fa',
+      principalId: '64242f8f-67ee-4f0c-b644-0766b42e8e91',
+      principalType: 'ServicePrincipal',
     };
 
-    const from: Entity = {
-      _class: ['AccessRole', 'AccessPolicy'],
-      _key: 'azure_role_definition-1',
-      _type: 'azure_role_definition',
-    };
-
-    const to: Entity = {
-      _class: 'User',
-      _key: 'azure_user-1',
-      _type: 'azure_user',
-    };
-
-    expect(
-      createRoleAssignmentDirectRelationship({
-        webLinker,
-        roleAssignment,
-        from,
-        to,
-      }),
-    ).toEqual({
-      _class: 'ASSIGNED',
-      _fromEntityKey: 'azure_role_definition-1',
-      _key: 'azure_role_definition-1|assigned|azure_user-1',
-      _toEntityKey: 'azure_user-1',
-      _type: 'azure_role_definition_assigned_user',
-      displayName: 'ASSIGNED',
-      id:
-        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleAssignments/c967042a-aad6-4e3b-8485-1c85d5e6f9e8',
-      name: 'c967042a-aad6-4e3b-8485-1c85d5e6f9e8',
-      principalId: 'b6e7f627-8731-47bd-be0e-477d1cbc6e17',
-      principalType: 'User',
-      roleDefinitionId:
-        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleDefinitions/8e3af657-a8ff-443c-a75c-2fe8c4bcb635',
-      scope: '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7',
-      type: 'Microsoft.Authorization/roleAssignments',
-      webLink:
-        'https://portal.azure.com/#@something.onmicrosoft.com/resource/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleAssignments/c967042a-aad6-4e3b-8485-1c85d5e6f9e8',
-    });
-  });
-});
-
-describe('createRoleAssignmentMappedRelationship', () => {
-  test('properties transferred', () => {
-    const roleAssignment: RoleAssignment = {
-      id:
-        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleAssignments/c967042a-aad6-4e3b-8485-1c85d5e6f9e8',
-      name: 'c967042a-aad6-4e3b-8485-1c85d5e6f9e8',
-      type: 'Microsoft.Authorization/roleAssignments',
-      scope: '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7',
-      roleDefinitionId:
-        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleDefinitions/8e3af657-a8ff-443c-a75c-2fe8c4bcb635',
-      principalId: 'b6e7f627-8731-47bd-be0e-477d1cbc6e17',
-      principalType: 'User',
-    };
-
-    const source: Entity = {
-      _class: ['AccessRole', 'AccessPolicy'],
-      _key: 'azure_role_definition-1',
-      _type: 'azure_role_definition',
-    };
-
-    const target = {
-      _key: 'azure_user-2',
-      _type: 'azure_user',
-    };
-
-    expect(
-      createRoleAssignmentMappedRelationship({
-        webLinker,
-        roleAssignment,
-        source,
-        target,
-      }),
-    ).toEqual({
-      _class: 'ASSIGNED',
-      _key: 'azure_role_definition-1|assigned|azure_user-2',
-      _mapping: {
-        relationshipDirection: 'FORWARD',
-        sourceEntityKey: 'azure_role_definition-1',
-        targetEntity: {
-          _key: 'azure_user-2',
-          _type: 'azure_user',
+    expect(createRoleAssignmentEntity(webLinker, data)).toEqual({
+      _class: ['AccessRole'],
+      _key:
+        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleAssignments/10000000-0000-0000-0000-000000000000',
+      _rawData: [
+        {
+          name: 'default',
+          rawData: {
+            id:
+              '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleAssignments/10000000-0000-0000-0000-000000000000',
+            name: '10000000-0000-0000-0000-000000000000',
+            principalId: '64242f8f-67ee-4f0c-b644-0766b42e8e91',
+            principalType: 'ServicePrincipal',
+            roleDefinitionId:
+              '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleDefinitions/8bfcfc94-cf28-d595-8e3d-851a1eb7c8fa',
+            scope: '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7',
+            type: 'Microsoft.Authorization/roleAssignments',
+          },
         },
-        targetFilterKeys: [['_type', '_key']],
-      },
-      _type: 'mapping_source_assigned_azure_user',
-      displayName: 'ASSIGNED',
+      ],
+      _type: 'azure_role_assignment',
+      canDelegate: undefined,
       id:
-        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleAssignments/c967042a-aad6-4e3b-8485-1c85d5e6f9e8',
-      name: 'c967042a-aad6-4e3b-8485-1c85d5e6f9e8',
-      principalId: 'b6e7f627-8731-47bd-be0e-477d1cbc6e17',
-      principalType: 'User',
+        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleAssignments/10000000-0000-0000-0000-000000000000',
+      name: '10000000-0000-0000-0000-000000000000',
+      displayName: '10000000-0000-0000-0000-000000000000',
+      principalId: '64242f8f-67ee-4f0c-b644-0766b42e8e91',
+      principalType: 'ServicePrincipal',
       roleDefinitionId:
-        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleDefinitions/8e3af657-a8ff-443c-a75c-2fe8c4bcb635',
+        '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleDefinitions/8bfcfc94-cf28-d595-8e3d-851a1eb7c8fa',
       scope: '/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7',
       type: 'Microsoft.Authorization/roleAssignments',
       webLink:
-        'https://portal.azure.com/#@something.onmicrosoft.com/resource/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleAssignments/c967042a-aad6-4e3b-8485-1c85d5e6f9e8',
+        'https://portal.azure.com/#@something.onmicrosoft.com/resource/subscriptions/d3803fd6-2ba4-4286-80aa-f3d613ad59a7/providers/Microsoft.Authorization/roleAssignments/10000000-0000-0000-0000-000000000000',
+      createdOn: undefined,
     });
   });
 });